Working on first Kaggle dataset
Explored the statistics of games that passed the minimum threshold in total purchases and total play hours for both levels of criteria. The main finding at the moment seems to be a tendency of games to follow a skewed distribution where the mean values are around the 75% quantile value for purchases and conversion rates and much higher for the number of hours played.